+++ packages/solid-crs-manage/lib/features/loan/models/loan-request-creation-args.ts
@@ -1,0 +1,4 @@
+export interface LoanRequestCreationArgs {
+  description: string;
+  collection: string;
+}