+++ packages/solid-crs-id-proxy/CHANGELOG.md
@@ -3,7 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### [0.16.3](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.16.3) (2022-04-22)
+## [0.17.0-loans.21](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.21) (2022-04-20)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.20](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.20) (2022-03-30)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.19](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.19) (2022-03-15)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.18](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.18) (2022-03-15)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.17](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.17) (2022-03-09)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.16](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.16) (2022-03-09)
+
+**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy
+
+
+
+
+
+## [0.17.0-loans.15](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.15) (2022-03-09)
 
 **Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-id-proxy