+++ packages/solid-crs-semcom-components/CHANGELOG.md
@@ -3,11 +3,102 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### [0.16.3](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.16.3) (2022-04-22)
-
-**Note:** Version bump only for package @netwerk-digitaal-erfgoed/solid-crs-semcom-components
+## [0.17.0-loans.21](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.21) (2022-04-20)
 
 
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+* implement LDN for metadata updates ([#598](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/598)) ([c4e0e4d](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/c4e0e4db5fa12d4a905cef9f9253cc6e8d6477e9))
+
+
+
+## [0.17.0-loans.20](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.20) (2022-03-30)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+
+
+
+## [0.17.0-loans.19](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.19) (2022-03-15)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+
+
+
+## [0.17.0-loans.18](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.18) (2022-03-15)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+
+
+
+## [0.17.0-loans.17](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.17) (2022-03-09)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+
+
+
+## [0.17.0-loans.16](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.16) (2022-03-09)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))
+
+
+
+## [0.17.0-loans.15](https://github.com/netwerk-digitaal-erfgoed/solid-crs/compare/v0.16.2...v0.17.0-loans.15) (2022-03-09)
+
+
+### **Bug Fixes**
+
+* set package versions to loans WIP ([#583](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/583)) ([48edbf8](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/48edbf8fb7bd2f0bb601dbde49ee887d53695c4a))
+
+
+### **Features**
+
+* create semcom object-loan component ([#578](https://github.com/netwerk-digitaal-erfgoed/solid-crs/issues/578)) ([a42dda2](https://github.com/netwerk-digitaal-erfgoed/solid-crs/commit/a42dda278ec963da6103ec4d6bbdb4d017bbf7aa))